Added ability to store tests in S3 using a run number instead of a timestamp which causes problems due to slight shift in time between propagation and runs.

+++ stack/corepersistence/perftest/src/main/java/org/apache/usergrid/perftest/NoopPerftest.java
@@ -30,7 +30,7 @@ public class NoopPerftest implements Perftest {
 
 
     @Override
-    public int getCallCount() {
+    public long getCallCount() {
         return 1000;
     }
 
@@ -42,7 +42,7 @@ public class NoopPerftest implements Perftest {
 
 
     @Override
-    public int getDelayBetweenCalls() {
+    public long getDelayBetweenCalls() {
         return 0;
     }
